+++ packages/components/src/Box/styles/Box.css.d.ts
@@ -1,0 +1,6 @@
+declare const styles: {
+  readonly "box": string;
+  readonly "preserve-white-space": string;
+};
+export = styles;
+