+++ packages/components/src/Box/styles/BoxBorder.css.d.ts
@@ -1,0 +1,10 @@
+declare const styles: {
+  readonly "border": string;
+  readonly "border-base": string;
+  readonly "border-top": string;
+  readonly "border-right": string;
+  readonly "border-bottom": string;
+  readonly "border-left": string;
+};
+export = styles;
+