+++ packages/components/src/Gallery/Gallery.test.tsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { fireEvent, render, waitFor } from "@testing-library/react";
+import * as browserUtilities from "@jobber/components/utils/getClientBrowser";
 import { Gallery } from ".";
 
 const files = [
@@ -9,6 +10,7 @@ const files = [
     type: "image/png",
     size: 213402324,
     progress: 1,
+    thumbnailSrc: "https://source.unsplash.com/50x50",
     src: "https://source.unsplash.com/250x250",
   },
   {
@@ -17,6 +19,7 @@ const files = [
     type: "image/png",
     size: 124525234,
     progress: 1,
+    thumbnailSrc: "https://source.unsplash.com/50x50",
     src: "https://source.unsplash.com/250x250",
   },
   {
@@ -25,6 +28,7 @@ const files = [
     type: "image/png",
     size: 233411234,
     progress: 1,
+    thumbnailSrc: "https://source.unsplash.com/50x50",
     src: "https://source.unsplash.com/250x250",
   },
   {
@@ -33,6 +37,7 @@ const files = [
     type: "image/png",
     size: 233411234,
     progress: 1,
+    thumbnailSrc: "https://source.unsplash.com/50x50",
     src: "https://source.unsplash.com/250x250",
   },
   {
@@ -41,6 +46,7 @@ const files = [
     type: "image/png",
     size: 233411234,
     progress: 1,
+    thumbnailSrc: "https://source.unsplash.com/50x50",
     src: "https://source.unsplash.com/250x250",
   },
 ];
@@ -166,4 +172,66 @@ describe("when a non-image is clicked", () => {
       expect(window.open).toHaveBeenCalledWith(pdfSrc, "_blank");
     });
   });
+});
+
+it("renders the appropriate thumbnail(icon) for an HEIC image not in Safari", async () => {
+  const heicFile = {
+    key: "123",
+    name: "sample.heic",
+    type: "image/heic",
+    size: 5000,
+    progress: 1,
+    src: "https://source.unsplash.com/250x250",
+  };
+  const { findByTestId } = render(<Gallery files={[heicFile]} />);
+
+  expect(await findByTestId("image")).toBeDefined();
+});
+
+it("renders the HEIC image thumbnail in Safari", async () => {
+  jest.spyOn(browserUtilities, "isSafari").mockReturnValue(true);
+  const heicFile = {
+    key: "123",
+    name: "sample.heic",
+    type: "image/heic",
+    size: 5000,
+    progress: 1,
+    src: "https://source.unsplash.com/250x250",
+  };
+  const { findByAltText } = render(<Gallery files={[heicFile]} />);
+  const imageElement = await findByAltText("sample.heic");
+  expect(imageElement).toHaveAttribute(
+    "src",
+    "https://source.unsplash.com/250x250",
+  );
+});
+
+describe("Thumbnails", () => {
+  it.each(files.map(file => [file.name, file.thumbnailSrc]))(
+    "should use the thumbnailSrc as the image source for %s",
+    async (fileName, src) => {
+      const { getByAltText } = render(<Gallery files={files} />);
+      const thumbnailImage = getByAltText(fileName);
+
+      await waitFor(() => {
+        expect(thumbnailImage).toHaveAttribute("src", src);
+      });
+    },
+  );
+
+  it.each(files.map(file => [file.name, file.src]))(
+    "should use the src as image source for %s when thumbnailSrc is not provided",
+    async (fileName, src) => {
+      const { getByAltText } = render(
+        <Gallery
+          files={files.map(file => ({ ...file, thumbnailSrc: undefined }))}
+        />,
+      );
+      const thumbnailImage = getByAltText(fileName);
+
+      await waitFor(() => {
+        expect(thumbnailImage).toHaveAttribute("src", src);
+      });
+    },
+  );
 });