+++ packages/components-native/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.69.1](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.69.0...@jobber/components-native@0.69.1) (2024-07-08)
+
+
+### Reverts
+
+* "feat: Storybook7 Upgrade" ([#1954](https://github.com/GetJobber/atlantis/issues/1954)) ([0854202](https://github.com/GetJobber/atlantis/commit/0854202ae850439c514f5ce1f129f2e584fe0efb)), closes [#1711](https://github.com/GetJobber/atlantis/issues/1711)
+
+
+
+
+
+# [0.69.0](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.68.0...@jobber/components-native@0.69.0) (2024-07-05)
+
+
+### Features
+
+* Storybook7 Upgrade ([#1711](https://github.com/GetJobber/atlantis/issues/1711)) ([7923df3](https://github.com/GetJobber/atlantis/commit/7923df325cae63a5476d997f577efd3e123dbe54))
+
+
+
+
+
+# [0.68.0](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.67.6...@jobber/components-native@0.68.0) (2024-06-24)
+
+
+### Features
+
+* **components-native:** Add testID prop to components-native Icon ([#1943](https://github.com/GetJobber/atlantis/issues/1943)) ([85faba9](https://github.com/GetJobber/atlantis/commit/85faba97f811420792ebe5bf3b91bbf8c79b2c19))
+
+
+
+
+
+## [0.67.6](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.67.5...@jobber/components-native@0.67.6) (2024-06-11)
+
+**Note:** Version bump only for package @jobber/components-native
+
+
+
+
+
 ## [0.67.5](https://github.com/GetJobber/atlantis/compare/@jobber/components-native@0.67.4...@jobber/components-native@0.67.5) (2024-06-03)
 
 **Note:** Version bump only for package @jobber/components-native